requirements.txt
bidict
#==0.22.1
Cartopy
#==0.21.1
certifi
#==2022.12.7
cftime
#==1.6.2
charset-normalizer
#==3.1.0
click
#==8.1.3
cmaps
#==1.0.5
contourpy
#==1.0.7
cycler
#==0.11.0
Flask
#==2.2.3
Flask-SocketIO
#==5.3.3
fonttools
#==4.39.2
geocat.viz
#==2023.3.0.post0
geos
#==0.2.3
h5netcdf
#==1.1.0
h5py
#==3.8.0
idna
#==3.4
itsdangerous
#==2.1.2
Jinja2
#==3.1.2
joblib
#==1.2.0
kiwisolver
#==1.4.4
lxml
#==4.9.2
MarkupSafe
#==2.1.2
matplotlib
#==3.7.1
MetPy
#==1.4.1
netCDF4
#==1.6.3
numpy
#==2.1.2
packaging
#==23.0
pandas
#==1.5.3
Pillow
#==9.4.0
Pint
#==0.20.1
platformdirs
#==3.2.0
pooch
#==1.7.0
pyparsing
#==3.0.9
pyproj
#==3.7.0
pyshp
#==2.3.1
python-dateutil
#==2.8.2
python-engineio
#==4.4.0
python-socketio
#==5.8.0
pytz
#==2023.2
requests
#==2.28.2
scikit-learn
#==1.2.2
scipy
#==1.10.1
shapely
#==2.0.1
six
#==1.16.0
threadpoolctl
#==3.1.0
traitlets
#==5.9.0
urllib3
#==1.26.15
Werkzeug
#==2.2.3
xarray
#==2023.3.0
gunicorn
#==20.1.0